requirements.txt
streamlit>=1.37
requests>=2.28
beautifulsoup4>=4.12
pdfplumber>=0.11
PyPDF2>=3.0
python-docx>=1.0.0
